% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calculate_new_pop_allele_freqs}
\alias{calculate_new_pop_allele_freqs}
\title{Calculate population allele frequencies based on new ancestor population frequencies}
\usage{
calculate_new_pop_allele_freqs(simulated_pop, ancestral_pop_seqs)
}
\arguments{
\item{simulated_pop}{The simulated population}

\item{ancestral_pop_seqs}{the ancestral population, should have the columns of ancestral_genotype, target_name, and seq}
}
\value{
population allele frequencies table with column names of target_name, seq, pop_allele_freq
}
\description{
Calculate population allele frequencies based on new ancestor population frequencies
}
\examples{
example_ancestors <- tibble(
ancestral_genotype = rep(paste0("AG", 1:4), times = 2),
target_name        = rep(c("target_A", "target_B"), each = 4),
seq = c(
  "AGCTAGCTAG",
  "CGTACGTACG",
  "TTGCAAGCTA",
  "GGCATTCGAA",
  "TACGATCGAT",
  "CCGGAATTCC",
  "ATATCGCGTA",
  "GGTTCCAAGC"
)
)
pop1 = sim_population(unique(example_ancestors$ancestral_genotype), 5, pop_alpha = 9, coi_r = 0.25, coi_p = 0.7, k_s = 0.5)
pop_freqs = calculate_new_pop_allele_freqs(pop1, example_ancestors)
}
