% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{draw_PLAF}
\alias{draw_PLAF}
\title{TODO.}
\usage{
draw_PLAF(df_loci, beta_shape1 = 10, beta_shape2 = 10)
}
\arguments{
\item{df_loci}{TODO}

\item{beta_shape1}{TODO}

\item{beta_shape2}{TODO}
}
\description{
TODO.
}
