% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dztnbinom}
\alias{dztnbinom}
\title{Zero-truncated negative binomial distribution}
\usage{
dztnbinom(x, size, prob, log = FALSE)
}
\arguments{
\item{x, size, prob, log}{these parameters are passed directly to `dnbinom()`.
See `?dnbinom` for further details of these parameters.}
}
\description{
Implements the zero-truncated negative binomial distribution, in
  which there is zero chance of seeing `n=0`.
}
\details{
The negative binomial distribution has several alternative
  parameterizations. We follow the same parameterization implemented in
  `dnbinom()`, with the small caveat that defining via `mu` is not available,
  only via `size` and `prob`. Note that the mean of the truncated
  distribution is different from the mean of the un-truncated distribution,
  now being given by `size*(1 - prob) / (prob*(1 - prob^size))`.
}
