% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_coi}
\alias{generate_coi}
\title{Generate a COI}
\usage{
generate_coi(coi_r, coi_p, max_coi = 100)
}
\arguments{
\item{coi_r, coi_p}{the r and p parameters to be given to the zero truncated negative binomial distribution COI generator `recombuddy::rztnbinom()` see `?rztnbinom` for more details}

\item{max_coi}{the maximum allowable COI}
}
\value{
a COI
}
\description{
generate COI
}
