% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_meiosis_number}
\alias{generate_meiosis_number}
\title{Generate the number of meiosis events}
\usage{
generate_meiosis_number(k_s, max_k = 20, off_set = 0)
}
\arguments{
\item{k_s}{the s parameter to be given to the type 1 geometric distribution random generator function `rgeom()` to select for serial meiosis}

\item{max_k}{the maximum k allowed}

\item{off_set}{off set to add, use 1 to generate a type 1 geometric distribution}
}
\value{
a number of meiosis events
}
\description{
generate the number of meiosis events with a cap on possible events and an offset (to allow Type 1 geometric distribution)
}
