% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_serial_meiosis_k}
\alias{generate_serial_meiosis_k}
\title{Generate a vector of recombinant parameters}
\usage{
generate_serial_meiosis_k(coi_r, coi_p, k_s, max_coi = 100)
}
\arguments{
\item{coi_r, coi_p}{the r and p parameters to be given to the zero truncated negative binomial distribution COI generator `recombuddy::rztnbinom()` see `?rztnbinom` for more details}

\item{k_s}{the s parameter to be given to the type 1 geometric distribution random generator function `rgeom()` to select for serial meiosis}

\item{max_coi}{the maximum allowable COI}
}
\value{
a vector of ks, to be given to `recombuddy::sim_sample()`
}
\description{
Will generate a vector of Serial meiosis parameters, the length will be the COI of the sample
}
\examples{
generate_serial_meiosis_k(0.25, 0.7, 0.5)
}
