% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_all_genotypes}
\alias{get_all_genotypes}
\title{Collect all `segment` tables together}
\usage{
get_all_genotypes(sim)
}
\arguments{
\item{sim}{a single simulated sample, as produced by `sim_sample()`.}
}
\description{
For a given sample, extracts the `segments` element of each
  haploid genotype into a single tibble.
}
