% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_utils.R
\name{get_ancestry_composition}
\alias{get_ancestry_composition}
\title{Get ancestry composition}
\usage{
get_ancestry_composition(df_genotypes)
}
\arguments{
\item{df_genotypes}{a data.frame (or tibble) of ancestry segments, such as that produced
by `get_all_genotypes()` for individual samples (see `?get_all_genotypes()`),
or by `get_population_genotypes()` for simulated populations (see `?get_population_genotypes()`).}
}
\value{
Returns a tibble with the following columns:
\itemize{
  \item `sample_id`
  \item `genotype`
  \item `index`
  \item `ancestry_total_bp`
  \item `ancestry_prop`
}
}
\description{
Calculates the overall genomic proportion inherited from each ancestral index.
}
\details{
The input is first passed through an internal helper that appends a
`segment_length` column if one is not already present. Ancestry proportions
are then calculated for each haploid genotype within sample(s). The output of
`get_ancestry_composition` can then serve as input for the plotting helper
function `plot_ancestry_composition`.
}
\examples{
n_set <- 10
set_props <- rdirichlet_single(n_set, alpha = 9)
samp1 <- sim_sample(k = c(0, 1, 2), set_props = set_props)
df_all_genotypes <- get_all_genotypes(samp1)

# For a single simulated sample
df_ancestry_sample <- get_ancestry_composition(df_all_genotypes)
df_ancestry_sample

# for a simulated population
# df_ancestry_population <- get_ancestry_composition(df_population_genotypes)
# df_ancestry_population

}
