% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_chrom_sizes}
\alias{get_chrom_sizes}
\title{Get chromosome lengths for P.falciparum 3D7}
\usage{
get_chrom_sizes()
}
\description{
Utility function that returns a vector of chromosome lengths for
  all 14 chromosomes of the 3D7 genome.
}
