% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_pf3d7_chrom_sizes}
\alias{get_pf3d7_chrom_sizes}
\title{Get chromosome lengths for P.falciparum 3D7}
\usage{
get_pf3d7_chrom_sizes()
}
\value{
returns a named vector of the lengths of P.falciparum 3D7 chromosomes
}
\description{
Utility function that returns a vector of chromosome lengths for
  all 14 chromosomes of the 3D7 genome (version 2020-09-01)
}
