% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_utils.R
\name{get_population_genotypes}
\alias{get_population_genotypes}
\title{Collect all `segment` tables together for all samples in a population}
\usage{
get_population_genotypes(sim_pop)
}
\arguments{
\item{sim_pop}{a simulated population, as produced by `sim_population()` (see `?sim_population()`).}
}
\value{
Returns a tibble with the following columns:
\itemize{
  \item `sample_id`
  \item `genotype`
  \item `chrom`
  \item `start`
  \item `end`
  \item `index`
}
}
\description{
For a population of samples, extracts the `segments` element of each
  haploid genotype per sample into a single population tibble.
}
\details{
This is a wrapper around `get_all_genotypes()` that applies the same logic
across all simulated samples in a population and appends a `sample_id` column
to identify which sample each genotype belongs to.
}
\examples{
# simulate a population of 5 samples
pop1 <- sim_population(
  paste0("sample", seq(0,100,1)),
  5,
  pop_alpha = 9,
  coi_r = 0.25,
  coi_p = 0.7,
  k_s = 0.5
)

df_population_genotypes <- get_population_genotypes(pop1)
df_population_genotypes
}
