% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_pvp01_chrom_sizes}
\alias{get_pvp01_chrom_sizes}
\title{Get chromosome lengths for P.vivax P01}
\usage{
get_pvp01_chrom_sizes()
}
\value{
returns a named vector of the lengths of P.vivax P01 chromosomes
}
\description{
Utility function that returns a vector of chromosome lengths for
  all 14 chromosomes of the P01 genome (version 2020-09-01)
}
