% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{intersect_panel_with_sample_genotypes}
\alias{intersect_panel_with_sample_genotypes}
\title{Intersect a panel with the all genotypes from a simulated sample}
\usage{
intersect_panel_with_sample_genotypes(panel_locs, simulated_sample)
}
\arguments{
\item{panel_locs}{a data frame with 4 columns, chrom, start, end, target, chrom values must match the recombined segment chroms}

\item{simulated_sample}{a simulated generated from `recombuddy::sim_sample`}
}
\value{
a table with the genotype and sample index from the simulated sample corresponding to the panel of intersect
}
\description{
Intersect a panel with the all genotypes from a simulated sample
}
\examples{
samp1 <- sim_sample(k = c(0, 2, 4), rho = 7.4e-7, set_props = rdirichlet_single(10, alpha = 9))
panel <- tibble(chrom = c("Pf3D7_04_v3", "Pf3D7_05_v3"), start = c(748173, 958071), end = c(748361, 958206), target = c("dhfr-1", "dhps-1"))
intersect_panel_with_sample_genotypes(panel, samp1)
}
