% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{intersect_panel_with_segs}
\alias{intersect_panel_with_segs}
\title{Intersect target locations for a panel with the recombined segments}
\usage{
intersect_panel_with_segs(panel_locs, segs)
}
\arguments{
\item{panel_locs}{a data frame with 4 columns, chrom, start, end, target, chrom values must match the recombined segment chroms}

\item{segs}{the segments from one of the genotypes generated from `sim_sample`}
}
\value{
a table with the sample indexes from which each panel intersects this, this is the ancestral genotype for this target in the panel
}
\description{
Intersect target locations for a panel with the recombined segments
}
\examples{
samp1 <- sim_sample(k = c(0, 2, 4), rho = 7.4e-7, set_props = rdirichlet_single(10, alpha = 9))
panel <- tibble(chrom = c("Pf3D7_04_v3", "Pf3D7_05_v3"), start = c(748173, 958071), end = c(748361, 958206), target = c("dhfr-1", "dhps-1"))
intersect_panel_with_segs(panel, samp1$genotypes[[3]]$segments)
}
