% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_utils.R
\name{plot_ancestry_composition}
\alias{plot_ancestry_composition}
\title{Plot ancestry composition}
\usage{
plot_ancestry_composition(df_ancestry, by = c("genotype", "sample"))
}
\arguments{
\item{df_ancestry}{a data.frame (or tibble) of ancestry composition, as
produced by `get_ancestry_composition()` (see `?get_ancestry_composition()`).}

\item{by}{the hierarchical level at which to display ancestry composition.
Must be either `"genotype"` or `"sample"`. The default is `"genotype"`.}
}
\value{
Returns a `ggplot2` object.
}
\description{
Produces a horizontal stacked bar plot showing the ancestry composition of
simulated genotypes. The length of each colored bar segment
corresponds to the genomic proportion inherited from a given ancestral index.
Separate panels are shown for each sample.
}
\details{
Requires columns `sample_id`, `genotype`, `index` and `ancestry_prop`.
If `by = "genotype"`, genotype-level ancestry proportions
are plotted and separated into panels by sample. If `by = "sample"`,
ancestry composition is first summed across genotypes within each sample and
then plotted at the sample level.
}
\examples{
n_set <- 10
set_props <- rdirichlet_single(n_set, alpha = 9)
samp1 <- sim_sample(k = c(0, 1, 2), set_props = set_props)
df_all_genotypes <- get_all_genotypes(samp1)
# genotype-level ancestry composition
df_all_genotypes |>
  get_ancestry_composition() |>
  plot_ancestry_composition(by = "genotype")

# sample-level ancestry composition
# df_population_genotypes |>
#  get_ancestry_composition() |>
#  plot_ancestry_composition(by = "sample")

}
