% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_read_counts}
\alias{plot_read_counts}
\title{TODO.}
\usage{
plot_read_counts(df_counts, chrom_sizes = get_chrom_sizes())
}
\arguments{
\item{df_counts}{TODO}

\item{chrom_sizes}{TODO}
}
\description{
TODO.
}
