% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_utils.R
\name{plot_segment_lengths}
\alias{plot_segment_lengths}
\title{Plot ancestry segment lengths}
\usage{
plot_segment_lengths(df_genotypes)
}
\arguments{
\item{df_genotypes}{a data.frame (or tibble) of ancestry segments, such as that produced
by `get_all_genotypes()` for individual samples (see `?get_all_genotypes()`),
or by `get_population_genotypes()` for simulated populations (see `?get_population_genotypes()`).}
}
\value{
Returns a `ggplot2` object.
}
\description{
Produces a boxplot of ancestry segment lengths from a data frame of simulated ancestry
segments.
}
\details{
The input is first passed through an internal helper that appends a
`segment_length` column if one is not already present. Note that boxplots for
non-recombinant genotypes reflect chromosome sizes since there is only
one ancestral index for the entire genome.
}
\examples{
n_set <- 10
set_props <- rdirichlet_single(n_set, alpha = 9)
samp1 <- sim_sample(k = c(0, 1, 2), set_props = set_props)
df_all_genotypes <- get_all_genotypes(samp1)
df_all_genotypes |> plot_segment_lengths()

}
