% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Draw from symmetric Dirichlet distribution}
\usage{
rdirichlet(n, alpha = 1)
}
\arguments{
\item{n}{the number of dimensions ofthe Dirichlet distribution.}

\item{alpha}{the unscaled concentration parameter. The final concentration
parameter is \code{alpha / n}}
}
\description{
Utility function for drawing from a symmetric Dirichlet
  distribution, where all concentration parameters are equal to \code{alpha / n}.
}
