% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rztnbinom}
\alias{rztnbinom}
\title{Zero-truncated negative binomial distribution random number generator}
\usage{
rztnbinom(n, size, prob)
}
\arguments{
\item{n}{the number of random numbers to generate}

\item{size}{passed directly to `dnbinom()`.See `?dnbinom` for further details of these parameters.}

\item{prob}{passed directly to `dnbinom()`.See `?dnbinom` for further details of these parameters.}
}
\value{
`n` random numbers from a Zero-truncated negative binomial distribution
}
\description{
Zero-truncated negative binomial distribution random number generator
}
\examples{
rztnbinom(10, 0.25, 0.7)
}
