% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_population_co_transmission}
\alias{sim_population_co_transmission}
\title{Population simulator}
\usage{
sim_population_co_transmission(
  input_samples,
  n_samples_out,
  pop_alpha,
  coi_r,
  coi_p,
  n_mosquitos_lambda,
  mosquitos_dir_conc,
  k_s,
  max_k = 20,
  max_coi = 100,
  rho = 7.4e-07,
  chrom_sizes = get_pf3d7_chrom_sizes(),
  n_cores = NULL
)
}
\arguments{
\item{input_samples}{a character list of samples, this will be the ancestral genotypes for the simulated population}

\item{n_samples_out}{the number of samples to generate}

\item{pop_alpha}{the alpha parameter to set for the dirichlet function for the population proportions, lower alphas generate more closely related final populations}

\item{coi_r, coi_p}{the r and p parameters to be given to the zero truncated negative binomial distribution COI generator `recombuddy::rztnbinom()` see `?rztnbinom` for more details}

\item{n_mosquitos_lambda}{the lambda value to be given to \code{\link{assign_strain_sources}}
to simulate number of mosquito bites that lead to this infection}

\item{mosquitos_dir_conc}{the alpha concentration value to be given to \code{\link{assign_strain_sources}}
to control the skew of strains being contributed by each mosquito, lower values mean}

\item{k_s}{the s parameter to be given to the type 1 geometric distribution random generator function `rgeom()` to select for serial meiosis}

\item{max_k}{the maximum k allowed}

\item{max_coi}{the maximum allowable COI}

\item{rho}{the recombination rate}

\item{chrom_sizes}{a named vector of lengths of chromosome lengths}
}
\value{
a list the simulated population and look up tables of population proportions, sample indexes, and chromosome sizes
}
\description{
Simulate a number of samples of a new population from a previous ancestral population, tries to adjust within-sample relatedness for amount of co-transmission
}
\examples{
# simulate pop_alpha 9 (~10\% between sample relatedness), coi_r = 0.5, coi_p = 0.2 (COI mean of 3.62, 32\% proportion will be monoclonal), k_s = 0.2 (20\% of genotypes will be recombinant)
n_mosquitos_lambda = 3.0 (multiple contributing mosquitoes probable, high EIR), mosquitos_dir_conc = 2.5  (strains spread more evenly across)
pop1 = sim_population_co_transmission(paste0("sample", seq(0,100,1)), 5, pop_alpha = 9, coi_r = 0.5, coi_p = 0.2, n_mosquitos_lambda = 3, mosquitos_dir_conc = 2.5, k_s = 0.2)
}
