% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_recomb}
\alias{sim_recomb}
\title{TODO.}
\usage{
sim_recomb(k, rho, set_props, chrom_sizes)
}
\arguments{
\item{k}{TODO}

\item{rho}{TODO}

\item{set_props}{TODO}

\item{chrom_sizes}{TODO}
}
\description{
TODO.
}
