% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_recomb_given_parents}
\alias{sim_recomb_given_parents}
\title{Simulate segments in a recombining genotype with supplied parents}
\usage{
sim_recomb_given_parents(
  k,
  parents,
  rho = 7.4e-07,
  chrom_sizes,
  zero_based_positioning = T
)
}
\arguments{
\item{k}{the number of serial meiosis events.}

\item{parents}{the parents to recombine}

\item{rho}{the recombination rate (per-site, per-meiosis). By default uses
the value 7.4e-7 from Miles et al. (2016).}

\item{chrom_sizes}{lengths of each chromosome, default taken from `get_pf3d7_chrom_sizes()`
by default.}
}
\description{
Produces a tibble of segments along the genome produced by
  meiosis of several distinct ancestors.
}
\details{
`k` sets the number of serial meiosis events. The number of
  ancestors chosen from the sample set is equal to 2^k. The indexes of
  the parents are supplied and need to match 2^k, same index can be given
  mutliple times
}
\references{
Miles A, Iqbal Z, Vauterin P, Pearson R, Campino S, Theron M, Gould K, Mead D, Drury E, O'Brien J, et al. (2016).
*Indels, structural variation, and recombination drive genomic diversity in Plasmodium falciparum*.
Genome Research, 26(9), 1288–1299.
\doi{10.1101/gr.203711.115}
}
