% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{validate_panel_locs_df}
\alias{validate_panel_locs_df}
\title{Validate panel locations}
\usage{
validate_panel_locs_df(panel_locs)
}
\arguments{
\item{panel_locs}{a data frame with 4 columns, chrom (character), start (numeric), end(numeric), target(character)}
}
\value{
nothing
}
\description{
Run validator with rules to make sure that there 4 columns and the following types chrom (character), start (numeric), end(numeric), target(character), will error out if fails to validate
}
\examples{
panel <- tibble(chrom = c("Pf3D7_04_v3", "Pf3D7_05_v3"), start = c(748173, 958071), end = c(748361, 958206), target = c("dhfr-1", "dhps-1"))
validate_panel_locs_df(panel)
}
