% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{draw_read_counts}
\alias{draw_read_counts}
\title{TODO.}
\usage{
draw_read_counts(df_WSAF, depth = 100, overdisp = 0.01)
}
\arguments{
\item{df_WSAF}{TODO}

\item{depth}{TODO}

\item{overdisp}{TODO}
}
\description{
TODO.
}
