% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_read_counts}
\alias{plot_read_counts}
\title{Plots draws from the read counts model}
\usage{
plot_read_counts(df_counts, chrom_sizes = get_pf3d7_chrom_sizes())
}
\arguments{
\item{df_counts}{a data.frame (tibble) of read counts, as produced by
`draw_read_counts()`.}

\item{chrom_sizes}{lengths of each chromosome, default taken from `get_pf3d7_chrom_sizes()`
by default.}
}
\description{
Takes values produced by `draw_read_counts()` and produces a
  simple plot showing *estimated* WSAFs along the genome. This simple plot
  often contains patterns that demonstrate clear blocks of relatedness.
}
