% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_sample}
\alias{sim_sample}
\title{TODO.}
\usage{
sim_sample(k, rho, set_props, chrom_sizes = get_chrom_sizes())
}
\arguments{
\item{k}{TODO}

\item{rho}{TODO}

\item{set_props}{TODO}

\item{chrom_sizes}{TODO}
}
\description{
TODO.
}
