% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_sample_co_transmission}
\alias{sim_sample_co_transmission}
\title{Simulate a single sample}
\usage{
sim_sample_co_transmission(
  coi,
  n_mosquitos_lambda,
  mosquitos_dir_conc,
  k_s,
  max_k = 20,
  coi_r,
  coi_p,
  max_coi = 100,
  rho = 7.4e-07,
  set_props,
  chrom_sizes = get_pf3d7_chrom_sizes()
)
}
\arguments{
\item{coi}{the number of strains to simulate}

\item{n_mosquitos_lambda}{the lambda value to be given to \code{\link{assign_strain_sources}}
to simulate number of mosquito bites that lead to this infection}

\item{mosquitos_dir_conc}{the alpha concentration value to be given to \code{\link{assign_strain_sources}}
to control the skew of strains being contributed by each mosquito, lower values mean}

\item{k_s}{the s parameter to be given to the type 1 geometric distribution random generator function `rgeom()` to select for serial meiosis}

\item{max_k}{the maximum k allowed}

\item{coi_r, coi_p}{the r and p parameters to be given to the zero truncated negative binomial distribution COI generator `recombuddy::rztnbinom()` see `?rztnbinom` for more details}

\item{max_coi}{the maximum allowable COI}

\item{rho}{the recombination rate (per-site, per-meiosis). By default uses
the value 7.4e-7 from Miles et al. (2016).}

\item{set_props}{proportions of each of the members of the sample set.
Dictates the probability they are chosen as an ancestor.}

\item{chrom_sizes}{lengths of each chromosome, default taken from `get_pf3d7_chrom_sizes()`
by default.}
}
\value{
A list with one element per haploid genome. Within each element is
  another list with three elements; the first two specify whether the
  genotype was copied over without recombination, and if so, which ancestor
  was it copied from. The final element `segments` contains the richest
  information - a tibble giving segments along the genome and which memeber
  of the sample set is ancestral to each segment.
}
\description{
Simulates one or more haploid genotypes within a single sample
  by drawing ancestors from a sample set and applying recombination.
}
\references{
Miles A, Iqbal Z, Vauterin P, Pearson R, Campino S, Theron M, Gould K, Mead D, Drury E, O'Brien J, et al. (2016).
*Indels, structural variation, and recombination drive genomic diversity in Plasmodium falciparum*.
Genome Research, 26(9), 1288–1299.
\doi{10.1101/gr.203711.115}

Miles A, Iqbal Z, Vauterin P, Pearson R, Campino S, Theron M, Gould K, Mead D, Drury E, O'Brien J, et al. (2016).
*Indels, structural variation, and recombination drive genomic diversity in Plasmodium falciparum*.
Genome Research, 26(9), 1288–1299.
\doi{10.1101/gr.203711.115}
}
